+++ tests/backends/megatron/test_model_provider_vpp.py
@@ -256,7 +256,7 @@ def test_model_provider_uses_mixture_lora_for_multiple_experts(monkeypatch):
     model = object()
     config = object()
     calls = []
-    mixture_module = types.ModuleType("relax.backends.megatron.mixture_lora")
+    mixture_module = types.ModuleType("relax.backends.megatron.mixture_lora_modules")
 
     def build_mixture_lora_peft(received_config, dropout, vp_stage=None):
         calls.append((received_config, dropout, vp_stage))
@@ -265,7 +265,7 @@ def test_model_provider_uses_mixture_lora_for_multiple_experts(monkeypatch):
     mixture_module.build_mixture_lora_peft = build_mixture_lora_peft
     mixture_module.ensure_mixture_lora_recompute_inputs_grad = lambda model: None
     mixture_module.install_mixture_lora_checkpoint_context = lambda: None
-    monkeypatch.setitem(sys.modules, "relax.backends.megatron.mixture_lora", mixture_module)
+    monkeypatch.setitem(sys.modules, "relax.backends.megatron.mixture_lora_modules", mixture_module)
     monkeypatch.setattr(module, "build_mixture_lora_config", lambda args: config)
     monkeypatch.setattr(module, "build_lora_peft", lambda args: pytest.fail("single LoRA factory was called"))
     monkeypatch.setattr(module, "validate_and_count_mixture_lora_parameters", lambda model: (100, 20, 4, 124))
